+++ packages/polythene-core-button-group/src/index.js
@@ -1,0 +1,4 @@
+import * as coreButtonGroup from "./button-group";
+export { coreButtonGroup };
+
+export { default as vars } from "./vars";