+++ packages/polythene-core-split-button/src/index.js
@@ -1,4 +1,0 @@
-import * as coreSplitButton from "./split-button";
-export { coreSplitButton };
-
-export { default as vars } from "./vars";