+++ packages/polythene-css-classes/button-group.js
@@ -1,0 +1,7 @@
+
+export default {
+  component: "pe-button-group",
+
+  // states
+  separator: "pe-button-group--separator",
+};