+++ packages/polythene-css-classes/split-button.js
@@ -1,7 +1,0 @@
-
-export default {
-  component: "pe-split-button",
-
-  // states
-  separator: "pe-split-button--separator",
-};