+++ packages/polythene-mithril-split-button/src/index.js
@@ -1,9 +1,0 @@
-import { ViewComponent } from "polythene-mithril-base";
-import { coreSplitButton as core } from "polythene-core-split-button";
-
-export const SplitButton = ViewComponent(Object.assign(
-  {},
-  core
-));
-
-SplitButton.displayName = "SplitButton";